# backend/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Wielerpool API
  version: 1.0.0
  description: Versieerbaar contract voor de gedeelde Tour- en Vuelta-backend.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /rounds/{roundId}:
    parameters:
      - $ref: "#/components/parameters/RoundId"
    get:
      security: []
      operationId: getRound
      responses:
        "200":
          description: Rondeconfiguratie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /rounds/{roundId}/participants:
    parameters:
      - $ref: "#/components/parameters/RoundId"
    get:
      security: []
      operationId: listParticipants
      responses:
        "200":
          description: Deelnemers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Participant"
  /rounds/{roundId}/riders:
    parameters:
      - $ref: "#/components/parameters/RoundId"
    get:
      security: []
      operationId: listRiders
      responses:
        "200":
          description: Renners en actuele prijzen
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rider"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wielerpool_session
  parameters:
    RoundId:
      name: roundId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
...
  responses:
...
    NotFound:
      description: Niet gevonden
...
  schemas:
...
    Round:
      type: object
      required: [id, name, competition, year, status, currency, settings]
      properties:
        id:
          type: string
        name:
          type: string
        competition:
          type: string
        year:
          type: integer
        status:
          type: string
          enum: [setup, test, registration, active, finished, archived]
        currency:
          type: string
        theme:
          type: object
          additionalProperties: true
        settings:
          $ref: "#/components/schemas/RoundSettings"
    RoundSettings:
      type: object
      additionalProperties: true
      required: [stake, budget, stageCount, starterCount, reserveCount, prizePotSplit]
      properties:
        stake:
          type: number
          minimum: 0
        budget:
          type: number
          minimum: 0
        stageCount:
          type: integer
          minimum: 1
        starterCount:
          type: integer
          minimum: 1
        reserveCount:
          type: integer
          minimum: 0
        scoringDepth:
          type: object
          additionalProperties:
            type: integer
        prizeWeights:
          type: object
          additionalProperties: true
        prizePotSplit:
          $ref: "#/components/schemas/PrizePotSplit"
        exchangeWindows:
          type: array
          items:
            $ref: "#/components/schemas/ExchangeWindow"
    PrizePotSplit:
      type: object
      additionalProperties: false
      required: [final, daily]
      properties:
        final:
          type: number
          minimum: 0
          maximum: 100
        daily:
          type: number
          minimum: 0
          maximum: 100
    ExchangeWindow:
      type: object
      required: [label, afterStage]
      properties:
        label:
          type: string
        afterStage:
          type: integer
          minimum: 0
        from:
          type: string
          format: date-time
        until:
          type: string
          format: date-time
    Participant:
      type: object
      required: [id, roundId, displayName]
      properties:
        id:
          type: string
          format: uuid
        roundId:
          type: string
        displayName:
          type: string
        teamName:
          type: string
        colorPrimary:
          type: string
        colorSecondary:
          type: string
    Rider:
      type: object
      required: [id, roundId, name, displayName, price]
      properties:
        id:
          type: string
          format: uuid
        roundId:
          type: string
        name:
          type: string
        displayName:
          type: string
        cyclingTeam:
          type: string
        price:
          type: number
          minimum: 0
        metadata:
          type: object
          additionalProperties: true
...